[#30] Refactor: comment 관련 query 커스텀 훅으로 분리
